requirements.txt
dj-database-url==0.4.1
Django==1.11.3
easy-thumbnails==2.4.1
gunicorn==19.6.0
olefile==0.44
Pillow==4.2.1
psycopg2==2.6.2
pytz==2017.2
requests==2.9.1
whitenoise==2.0.6
